finish moving stuff and adding `ValidNext`
