+++ tbdex/offering/create.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/TBD54566975/tbdex-go/tbdex"
 	"github.com/TBD54566975/tbdex-go/tbdex/crypto"
+	"github.com/TBD54566975/tbdex-go/tbdex/resource"
 	"github.com/tbd54566975/web5-go/dids/did"
 	"github.com/tbd54566975/web5-go/pexv2"
 	"go.jetpack.io/typeid"
@@ -40,7 +40,7 @@ func Create(fromDID did.BearerDID, payin *PayinDetails, payout *PayoutDetails, r
 	}
 
 	offering := Offering{
-		Metadata: tbdex.ResourceMetadata{
+		Metadata: resource.Metadata{
 			From:      fromDID.URI,
 			Kind:      Kind,
 			ID:        o.id.String(),
